requirements.txt
streamlit>=1.48.0
fastmcp>=2.11.0
openai>=1.100.0
python-dotenv>=1.0.0
